+++ src/types.ts
@@ -4,7 +4,9 @@ export type Constructor<T extends any> = new (...args: any[]) => T;
 export type MethodCalls<Object> = Record<keyof Object, number>;
 
 export type MethodOf<Object> = {
-  [K in keyof Object]: Object[K] extends (...args: unknown[]) => unknown ? K : never;
+  // deno-lint-ignore no-explicit-any
+  [K in keyof Object]: Object[K] extends (...args: any[]) => unknown ? K
+    : never;
 }[keyof Object];
 
 export type MemberOf<Object> = {
@@ -13,5 +15,6 @@ export type MemberOf<Object> = {
 
 export type MockedObject = { [k: string]: unknown };
 
-export type StubReturnValue<T, R> = T extends (...args: unknown[]) => unknown ? () => R
+export type StubReturnValue<T, R> = T extends (...args: unknown[]) => unknown
+  ? () => R
   : string;
